refactor: move loadClock into `clock` file
